+++ interfaces/CitationInfoBook.ts
@@ -1,0 +1,6 @@
+export default interface CitationInfoBook {
+  authors: string | undefined;
+  date: string | undefined;
+  title: string | undefined;
+  editorial: string | undefined;
+}